plagiarism_checker/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
python-dotenv
PyPDF2
python-docx
scikit-learn
numpy